# docs/swagger.yaml
basePath: /shyft
definitions:
  handlers.RespondJson:
    properties:
      intent:
        type: string
      message: {}
      status:
        type: boolean
    type: object
  handlers.updateShiftScheduleDTO:
    properties:
      alias:
        type: string
      description:
        type: string
      end_date:
        type: string
      frequency:
        description: 1, 2, 3, 4, 5, 6, 7 (days of the week)
        type: integer
      manager:
        items: {}
        type: array
      organization:
        items: {}
        type: array
      shifts:
        items: {}
        type: array
      start_date:
        type: string
      status:
        description: '0: pending, 1: approved, 2: rejected'
        type: integer
      users:
        items: {}
        type: array
      year:
        type: integer
    required:
    - alias
    - description
    - end_date
    - frequency
    - manager
    - organization
    - shifts
    - start_date
    - users
    - year
    type: object
...
info:
  contact:
    email: YunusAlpu@icloud.com
    name: Yunus Emre Alpu
    url: https://yunusemrealpu.netlify.app
  title: Shift Scheduler Service API
  version: 1.0.0
paths:
  /shift-schedule/{id}:
    put:
      consumes:
      - application/json
      description: update a shift schedule
      parameters:
      - description: Shift Schedule ID
        in: path
        name: id
        required: true
        type: string
      - description: update shift schedule
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.updateShiftScheduleDTO'
      produces:
      - application/json
      responses:
        "200":
          description: successfully updated shift schedule
          schema:
            $ref: '#/definitions/handlers.RespondJson'
        "400":
          description: cannot update shift schedule due to invalid request body
          schema:
            $ref: '#/definitions/handlers.RespondJson'
        "422":
          description: cannot update shift schedule due to invalid request body
          schema:
            $ref: '#/definitions/handlers.RespondJson'
        "500":
          description: cannot update shift schedule due to internal server error
          schema:
            $ref: '#/definitions/handlers.RespondJson'
      security:
      - BearerAuth: []
      summary: update a shift schedule
      tags:
      - Shift
...
schemes:
- http
- https
swagger: "2.0"